Figured a way to test expired link can not be used

+++ src/main/java/com/rdi/urlshortener/services/RDIUrlShortenerService.java
@@ -5,6 +5,7 @@ import com.rdi.urlshortener.data.models.Url;
 import com.rdi.urlshortener.data.repositories.UrlShortenerRepository;
 import com.rdi.urlshortener.dto.responses.DeleteUrlResponse;
 import com.rdi.urlshortener.dto.responses.GenerateShortUrlResponse;
+import com.rdi.urlshortener.exception.UrlExpiredException;
 import com.rdi.urlshortener.exception.UrlNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,18 +29,46 @@ public class RDIUrlShortenerService implements UrlShortenerService{
         return map(savedUrl);
     }
 
-    @Override
-    public DeleteUrlResponse deleteUrl(Long urlId) throws UrlNotFoundException {
-        Url foundUrl = urlShortenerRepository.findById(urlId)
-                .orElseThrow(() -> new UrlNotFoundException("URL NOT FOUND"));
-        urlShortenerRepository.delete(foundUrl);
-
-        return new DeleteUrlResponse("URL DELETED SUCCESSFULLY");
-    }
-
     private static String hashOriginalUrl(String originalUrl) {
         return Hashing.murmur3_32_fixed()
                 .hashString(originalUrl.concat(String.valueOf(LocalDateTime.now())), StandardCharsets.UTF_8)
                 .toString();
     }
+
+    @Override
+    public DeleteUrlResponse deleteUrl(String shortUrl) throws UrlNotFoundException {
+        Url foundUrl = getUrl(shortUrl);
+        urlShortenerRepository.delete(foundUrl);
+        return new DeleteUrlResponse("URL DELETED SUCCESSFULLY");
+    }
+
+    @Override
+    public String shortUrlRedirect(String shortUrl) throws UrlExpiredException, UrlNotFoundException {
+        Url foundUrl = getUrl(shortUrl);
+        checkIfUrlIsExpired(foundUrl);
+        return foundUrl.getOriginalUrl();
+    }
+
+    private void checkIfUrlIsExpired(Url foundUrl) throws UrlExpiredException {
+        if(foundUrl.getExpirationDateAndTime().isBefore(LocalDateTime.now())){
+            urlShortenerRepository.delete(foundUrl);
+            throw new UrlExpiredException("THIS SHORT URL HAS EXPIRED.");
+        }
+    }
+
+    @Override
+    public void updateUrlExpiryDateJustForTesting(String shortUrl) throws UrlNotFoundException {
+        Url found = getUrl(shortUrl);
+        if (found.getOriginalUrl().equals("http://www.rdi.io")) {
+            found.setExpirationDateAndTime(LocalDateTime.now().minusDays(2));
+            urlShortenerRepository.save(found);
+        }
+    }
+
+    private Url getUrl(String shortUrl) throws UrlNotFoundException {
+        return urlShortenerRepository.findByShortUrl(shortUrl)
+                .orElseThrow(() -> new UrlNotFoundException("URL NOT FOUND"));
+    }
+
+
 }
